requirements.txt
ollama
Pillow
gradio

